feat: Build FLEX & STAX OK - Screens to build
